+++ config/manager/kustomization.yaml
@@ -3,6 +3,6 @@ resources:
 images:
 - name: gcr.io/datadoghq/watermarkpodautoscaler
   newName: gcr.io/datadoghq/watermarkpodautoscaler
-  newTag: v0.8.0-rc.2
+  newTag: v0.8.0-rc.2-5d4797e
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
